Relax URL validation
Relaxes the URL validation to just ensure it begins with http:// or https:// (ignoring case), and that it has at least 1 character after that, with no spaces.  This is mostly in line with the discussion of the short term fix for #21 

Note that since this currently only runs on the output of normalizeUrl, it will be guaranteed to start with either https:// or http://, which is why I left that validation in there anyway.